Split completion resolve / providing into separate endpoints.
- When playing with the completion stack I felt that resolve / providing was becoming a bit unweidly and concerns were bleeding across the domain. Therefore I went ahead and split the resolve / provide endpoints apart so the logic is kept separately. This is similar to how we have code actions today as well.
- Also split tests and updated accordingly.
- Ended up needing to promote our completion list cache into a service so both endpoints could interact with the same completion list cache without relying on member state.
