+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Completion/RazorCompletionEndpointTest.cs
@@ -11,7 +11,6 @@ using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.Components;
 using Microsoft.AspNetCore.Razor.LanguageServer.Common.Extensions;
 using Microsoft.AspNetCore.Razor.LanguageServer.ProjectSystem;
-using Microsoft.AspNetCore.Razor.LanguageServer.Tooltip;
 using Microsoft.AspNetCore.Razor.Test.Common;
 using Microsoft.CodeAnalysis.Razor.Completion;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
@@ -47,18 +46,15 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
                 new TagHelperCompletionProvider(tagHelperCompletionService, new DefaultHtmlFactsService(), tagHelperFactsService)
             };
             CompletionFactsService = new DefaultRazorCompletionFactsService(completionProviders);
-            LSPTagHelperTooltipFactory = Mock.Of<LSPTagHelperTooltipFactory>(MockBehavior.Strict);
-            VSLSPTagHelperTooltipFactory = Mock.Of<VSLSPTagHelperTooltipFactory>(MockBehavior.Strict);
             EmptyDocumentResolver = Mock.Of<DocumentResolver>(MockBehavior.Strict);
+            CompletionListCache = new CompletionListCache();
         }
 
         private RazorCompletionFactsService CompletionFactsService { get; }
 
-        private LSPTagHelperTooltipFactory LSPTagHelperTooltipFactory { get; }
-
-        private VSLSPTagHelperTooltipFactory VSLSPTagHelperTooltipFactory { get; }
-
         private DocumentResolver EmptyDocumentResolver { get; }
+
+        private CompletionListCache CompletionListCache { get; }
 
         [Fact]
         public void IsApplicableTriggerContext_Deletion_ReturnsFalse()
@@ -335,216 +331,6 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             Assert.Null(converted.Command);
         }
 
-        [Fact]
-        public async Task Handle_Resolve_DirectiveCompletion_ReturnsCompletionItemWithDocumentation()
-        {
-            // Arrange
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
-            var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, EmptyDocumentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
-            var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.Directive);
-            razorCompletionItem.SetDirectiveCompletionDescription(new DirectiveCompletionDescription("Test directive"));
-            var completionList = completionEndpoint.CreateLSPCompletionList(new[] { razorCompletionItem });
-            var completionItem = completionList.Items.Single();
-
-            // Act
-            var newCompletionItem = await completionEndpoint.Handle(completionItem, default);
-
-            // Assert
-            Assert.NotNull(newCompletionItem.Documentation);
-        }
-
-        [Fact]
-        public async Task Handle_Resolve_MarkupTransitionCompletion_ReturnsCompletionItemWithDocumentation()
-        {
-            // Arrange
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
-            var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, EmptyDocumentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
-            var razorCompletionItem = new RazorCompletionItem("@...", "@", RazorCompletionItemKind.MarkupTransition);
-            razorCompletionItem.SetMarkupTransitionCompletionDescription(new MarkupTransitionCompletionDescription("Test description"));
-            var completionList = completionEndpoint.CreateLSPCompletionList(new[] { razorCompletionItem });
-            var completionItem = completionList.Items.Single();
-
-            // Act
-            var newCompletionItem = await completionEndpoint.Handle(completionItem, default);
-
-            // Assert
-            Assert.NotNull(newCompletionItem.Documentation);
-        }
-
-        [Fact]
-        public async Task Handle_Resolve_DirectiveAttributeCompletion_ReturnsCompletionItemWithDocumentation()
-        {
-            // Arrange
-            var lspDescriptionFactory = new Mock<LSPTagHelperTooltipFactory>(MockBehavior.Strict);
-            var markdown = new MarkupContent
-            {
-                Kind = MarkupKind.Markdown,
-                Value = "Some Markdown"
-            };
-            lspDescriptionFactory.Setup(factory => factory.TryCreateTooltip(It.IsAny<AggregateBoundAttributeDescription>(), out markdown))
-                .Returns(true);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
-            var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, EmptyDocumentResolver, CompletionFactsService, lspDescriptionFactory.Object, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
-            var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.DirectiveAttribute);
-            razorCompletionItem.SetAttributeCompletionDescription(new AggregateBoundAttributeDescription(Array.Empty<BoundAttributeDescriptionInfo>()));
-            var completionList = completionEndpoint.CreateLSPCompletionList(new[] { razorCompletionItem });
-            var completionItem = completionList.Items.Single();
-
-            // Act
-            var newCompletionItem = await completionEndpoint.Handle(completionItem, default);
-
-            // Assert
-            Assert.NotNull(newCompletionItem.Documentation);
-        }
-
-        [Fact]
-        public async Task Handle_Resolve_DirectiveAttributeParameterCompletion_ReturnsCompletionItemWithDocumentation()
-        {
-            // Arrange
-            var descriptionFactory = new Mock<LSPTagHelperTooltipFactory>(MockBehavior.Strict);
-            var markdown = new MarkupContent
-            {
-                Kind = MarkupKind.Markdown,
-                Value = "Some Markdown"
-            };
-            descriptionFactory.Setup(factory => factory.TryCreateTooltip(It.IsAny<AggregateBoundAttributeDescription>(), out markdown))
-                .Returns(true);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
-            var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, EmptyDocumentResolver, CompletionFactsService, descriptionFactory.Object, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
-            var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.DirectiveAttributeParameter);
-            razorCompletionItem.SetAttributeCompletionDescription(new AggregateBoundAttributeDescription(Array.Empty<BoundAttributeDescriptionInfo>()));
-            var completionList = completionEndpoint.CreateLSPCompletionList(new[] { razorCompletionItem });
-            var completionItem = completionList.Items.Single();
-
-            // Act
-            var newCompletionItem = await completionEndpoint.Handle(completionItem, default);
-
-            // Assert
-            Assert.NotNull(newCompletionItem.Documentation);
-        }
-
-        [Fact]
-        public async Task Handle_Resolve_TagHelperElementCompletion_ReturnsCompletionItemWithDocumentation()
-        {
-            // Arrange
-            var lspDescriptionFactory = new Mock<LSPTagHelperTooltipFactory>(MockBehavior.Strict);
-            var markdown = new MarkupContent
-            {
-                Kind = MarkupKind.Markdown,
-                Value = "Some Markdown"
-            };
-            lspDescriptionFactory.Setup(factory => factory.TryCreateTooltip(It.IsAny<AggregateBoundElementDescription>(), out markdown))
-                .Returns(true);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
-            var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, EmptyDocumentResolver, CompletionFactsService, lspDescriptionFactory.Object, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
-            var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.TagHelperElement);
-            razorCompletionItem.SetTagHelperElementDescriptionInfo(new AggregateBoundElementDescription(Array.Empty<BoundElementDescriptionInfo>()));
-            var completionList = completionEndpoint.CreateLSPCompletionList(new[] { razorCompletionItem });
-            var completionItem = completionList.Items.Single();
-
-            // Act
-            var newCompletionItem = await completionEndpoint.Handle(completionItem, default);
-
-            // Assert
-            Assert.NotNull(newCompletionItem.Documentation);
-        }
-
-        [Fact]
-        public async Task Handle_Resolve_TagHelperElementCompletion_NullCommitCharacters()
-        {
-            // Arrange
-            var vsLSPDescriptionFactory = new Mock<VSLSPTagHelperTooltipFactory>(MockBehavior.Strict);
-            var markdown = new VSClassifiedTextElement(new VSClassifiedTextRun("type", "text"));
-            vsLSPDescriptionFactory.Setup(factory => factory.TryCreateTooltip(It.IsAny<AggregateBoundElementDescription>(), out markdown))
-                .Returns(true);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams()
-            {
-                Capabilities = new PlatformAgnosticClientCapabilities
-                {
-                    SupportsVisualStudioExtensions = true,
-                }
-            });
-            var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, EmptyDocumentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, vsLSPDescriptionFactory.Object, languageServer.Object, LoggerFactory);
-            var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.TagHelperElement);
-            razorCompletionItem.SetTagHelperElementDescriptionInfo(new AggregateBoundElementDescription(Array.Empty<BoundElementDescriptionInfo>()));
-            var completionList = completionEndpoint.CreateLSPCompletionList(new[] { razorCompletionItem });
-            var completionItem = completionList.Items.Single();
-
-            // For performance we sometimes return null commit characters, so we need to be sure they resolve safely;
-            completionItem = completionItem with { CommitCharacters = null };
-
-            // Act
-            var newCompletionItem = (VSCompletionItem)await completionEndpoint.Handle(completionItem, default);
-
-            // Assert
-            Assert.NotNull(newCompletionItem.Description);
-        }
-
-        [Fact]
-        public async Task Handle_Resolve_TagHelperAttribute_ReturnsCompletionItemWithDocumentation()
-        {
-            // Arrange
-            var lspDescriptionFactory = new Mock<LSPTagHelperTooltipFactory>(MockBehavior.Strict);
-            var markdown = new MarkupContent
-            {
-                Kind = MarkupKind.Markdown,
-                Value = "Some Markdown"
-            };
-            lspDescriptionFactory.Setup(factory => factory.TryCreateTooltip(It.IsAny<AggregateBoundAttributeDescription>(), out markdown))
-                .Returns(true);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
-            var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, EmptyDocumentResolver, CompletionFactsService, lspDescriptionFactory.Object, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
-            var razorCompletionItem = new RazorCompletionItem("TestItem", "TestItem", RazorCompletionItemKind.TagHelperAttribute);
-            razorCompletionItem.SetAttributeCompletionDescription(new AggregateBoundAttributeDescription(Array.Empty<BoundAttributeDescriptionInfo>()));
-            var completionList = completionEndpoint.CreateLSPCompletionList(new[] { razorCompletionItem });
-            var completionItem = completionList.Items.Single();
-
-            // Act
-            var newCompletionItem = await completionEndpoint.Handle(completionItem, default);
-
-            // Assert
-            Assert.NotNull(newCompletionItem.Documentation);
-        }
-
-        [Fact]
-        public async Task Handle_Resolve_NonTagHelperCompletion_Noops()
-        {
-            // Arrange
-            var lspDescriptionFactory = new Mock<LSPTagHelperTooltipFactory>(MockBehavior.Strict);
-            var markdown = new MarkupContent
-            {
-                Kind = MarkupKind.Markdown,
-                Value = "Some Markdown"
-            };
-            lspDescriptionFactory.Setup(factory => factory.TryCreateTooltip(It.IsAny<AggregateBoundElementDescription>(), out markdown))
-                .Returns(true);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
-            var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, EmptyDocumentResolver, CompletionFactsService, lspDescriptionFactory.Object, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
-            var completionItem = new CompletionItem();
-
-            // Act
-            var newCompletionItem = await completionEndpoint.Handle(completionItem, default);
-
-            // Assert
-            Assert.Null(newCompletionItem.Documentation);
-        }
-
         // This is more of an integration test to validate that all the pieces work together
         [Fact]
         public async Task Handle_Unsupported_NoCompletionItems()
@@ -554,10 +340,8 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             var codeDocument = CreateCodeDocument("@");
             codeDocument.SetUnsupported();
             var documentResolver = CreateDocumentResolver(documentPath, codeDocument);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
             var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, documentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
+                Dispatcher, documentResolver, CompletionFactsService, CompletionListCache, LoggerFactory);
             var request = new CompletionParams()
             {
                 TextDocument = new TextDocumentIdentifier(new Uri(documentPath)),
@@ -580,10 +364,8 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             var documentPath = "C:/path/to/document.cshtml";
             var codeDocument = CreateCodeDocument("@");
             var documentResolver = CreateDocumentResolver(documentPath, codeDocument);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
             var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, documentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
+                Dispatcher, documentResolver, CompletionFactsService, CompletionListCache, LoggerFactory);
             var request = new CompletionParams()
             {
                 TextDocument = new TextDocumentIdentifier(new Uri(documentPath)),
@@ -615,10 +397,8 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             var codeDocument = CreateCodeDocument("@in");
             codeDocument.SetTagHelperContext(tagHelperContext);
             var documentResolver = CreateDocumentResolver(documentPath, codeDocument);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
             var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, documentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
+                Dispatcher, documentResolver, CompletionFactsService, CompletionListCache, LoggerFactory);
             var request = new CompletionParams()
             {
                 TextDocument = new TextDocumentIdentifier(new Uri(documentPath)),
@@ -649,10 +429,8 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             var codeDocument = CreateCodeDocument("@inje");
             codeDocument.SetTagHelperContext(tagHelperContext);
             var documentResolver = CreateDocumentResolver(documentPath, codeDocument);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
             var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, documentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
+                Dispatcher, documentResolver, CompletionFactsService, CompletionListCache, LoggerFactory);
             var request = new CompletionParams()
             {
                 TextDocument = new TextDocumentIdentifier(new Uri(documentPath)),
@@ -683,10 +461,8 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             var codeDocument = CreateCodeDocument("@inje");
             codeDocument.SetTagHelperContext(tagHelperContext);
             var documentResolver = CreateDocumentResolver(documentPath, codeDocument);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
             var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, documentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
+                Dispatcher, documentResolver, CompletionFactsService, CompletionListCache, LoggerFactory);
             var request = new CompletionParams()
             {
                 TextDocument = new TextDocumentIdentifier(new Uri(documentPath)),
@@ -718,10 +494,8 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             var codeDocument = CreateCodeDocument("<");
             codeDocument.SetTagHelperContext(tagHelperContext);
             var documentResolver = CreateDocumentResolver(documentPath, codeDocument);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
             var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, documentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
+                Dispatcher, documentResolver, CompletionFactsService, CompletionListCache, LoggerFactory);
             var request = new CompletionParams()
             {
                 TextDocument = new TextDocumentIdentifier(new Uri(documentPath)),
@@ -756,10 +530,8 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             var codeDocument = CreateCodeDocument("<test  ");
             codeDocument.SetTagHelperContext(tagHelperContext);
             var documentResolver = CreateDocumentResolver(documentPath, codeDocument);
-            var languageServer = new Mock<ClientNotifierServiceBase>(MockBehavior.Strict);
-            languageServer.Setup(ls => ls.ClientSettings).Returns(new InitializeParams());
             var completionEndpoint = new RazorCompletionEndpoint(
-                Dispatcher, documentResolver, CompletionFactsService, LSPTagHelperTooltipFactory, VSLSPTagHelperTooltipFactory, languageServer.Object, LoggerFactory);
+                Dispatcher, documentResolver, CompletionFactsService, CompletionListCache, LoggerFactory);
             var request = new CompletionParams()
             {
                 TextDocument = new TextDocumentIdentifier(new Uri(documentPath)),
